+++ moabb/tests/contexts.py
@@ -1,7 +1,7 @@
 from moabb.evaluations import evaluations as ev
 from moabb.datasets.bnci import BNCI2014001
 from moabb.analysis import Results
-from moabb.contexts.motor_imagery import LeftRightImagery
+from moabb.paradigms.motor_imagery import LeftRightImagery
 import unittest
 
 from pyriemann.spatialfilters import CSP
